refactor: extract OG asset plugins into src/plugins/og-assets.ts
Move vinext:og-inline-fetch-assets and vinext:og-assets out of the
monolithic index.ts into a focused file (src/plugins/og-assets.ts).

- createOgInlineFetchAssetsPlugin() factory: owns the per-build cache
  (Map<string, string>) and isBuild flag as closure state, removing
  the two module-level variables that leaked into the outer vinext()
  factory closure.
- ogAssetsPlugin constant: the writeBundle-only plugin has no mutable
  state, so it exports directly as a plain plugin object.

index.ts now imports and delegates to these, removing ~190 lines of
inline plugin code from the main plugin file.
